Some small edits
Absolutely exhausted rn so leaving this for the morning. Nav bar is not sticky and I can't figure out why rn, but I think that could ultimately be a cool feature. If it doesn't work, though, we should make  left-side small column with options of places to jump? or we can just force people to scroll. kinda like jumping though. CSS mentioned earlier still not fixed, but will do later.